+++ src/mime.rs
@@ -1,5 +1,5 @@
 /// List of allowed mimes.
-static ALLOWED_MIME_TYPES: [&str; 2] = ["text/plain;charset=utf-8", "UTF8_STRING"];
+static ALLOWED_MIME_TYPES: [&str; 3] = ["text/plain;charset=utf-8", "text/plain", "UTF8_STRING"];
 
 /// Mime type supported by clipboard.
 #[derive(Clone, Copy, Eq, PartialEq, Debug)]
@@ -8,11 +8,15 @@ pub enum MimeType {
     ///
     /// The primary mime type used by most clients
     TextPlainUtf8 = 0,
+    /// text/plain mime type.
+    ///
+    /// Fallback without charset parameter
+    TextPlain = 1,
     /// UTF8_STRING mime type.
     ///
     /// Some X11 clients are using only this mime type, so we
     /// should have it as a fallback just in case.
-    Utf8String = 1,
+    Utf8String = 2,
 }
 
 impl MimeType {
@@ -24,6 +28,8 @@ impl MimeType {
         for offered_mime_type in offered_mime_types.iter() {
             if offered_mime_type == ALLOWED_MIME_TYPES[Self::TextPlainUtf8 as usize] {
                 return Some(Self::TextPlainUtf8);
+            } else if offered_mime_type == ALLOWED_MIME_TYPES[Self::TextPlain as usize] {
+                return Some(Self::TextPlain);
             } else if offered_mime_type == ALLOWED_MIME_TYPES[Self::Utf8String as usize] {
                 return Some(Self::Utf8String);
             }